+++ tests/testthat/test-ts_first_of_period.R
@@ -7,6 +7,7 @@ test_that("ts_first_of_period works", {
   )
   ans <- ts_first_of_period(x)
   expect_true(all(data.table::mday(ans$time) == 1))
+  expect_identical(nrow(ans), nrow(x))
 
   ans <- ts_first_of_period(ts_lag(ts_df(austres), "14 days"))
   expect_true(all(data.table::mday(ans$time) == 1))
@@ -16,6 +17,16 @@ test_that("ts_first_of_period works", {
     value = rnorm(10)
   ), "3 sec")
   ans <- ts_first_of_period(x)
+  expect_identical(nrow(ans), nrow(x))
 
   expect_true(all(as.integer(ans$time) %% 10 == 0))
+})
+
+
+test_that("ts_first_of_period works with POSIXct #210", {
+  x <- ts_lag(data.frame(
+    time = seq(as.POSIXct("1970-01-01"), length.out = 10, by = "10 sec"),
+    value = rnorm(10)
+  ), "3 sec")
+  expect_true(nrow(ts_first_of_period(x)) == 10)
 })